--- logger2/src/test/scala/com/tersesystems/echopraxia/logger2/Models.scala
@@ -1,7 +1,0 @@
-package com.tersesystems.echopraxia.logger2
-
-
-case class Book(category: String, author: String, title: String, price: Double)
-
-case class CreditCard(number: String, expirationDate: String)
-
